Evolution tests covering more cases. Part of #10

+++ tests/dom2/dom/CaModel_scriptNoUpdate.lua
@@ -27,7 +27,7 @@ Type "dom.ICompany"
 Type "dom.IEmployee"
 {
 	name = "string",
-	salary = "double",
+	salary = "int32",
 	role = "string",
 }
